+++ release_notes.md
@@ -1,4 +1,5 @@
 ### 0.2.0 _ 2016-02 under construction
+  * [#6](https://github.com/ctaggart/froto/issues/6) Add support for proto3 language
   * [#8](https://github.com/ctaggart/froto/issues/8) Strict/complete proto2 compliance
   * [#9](https://github.com/ctaggart/froto/issues/9) updated build & dependencies
     * switched from Apache 2 to MIT license